requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.7
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.12
langchain-core==0.2.29
langchain-google-genai==1.0.8
langchain-text-splitters==0.2.2
langsmith==0.1.98
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.32
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
yarl==1.9.4
